backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-richtextfield==1.6
djangorestframework==3.12.4
Pillow==8.1.2
PyJWT==1.7.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
